# src/main/java/hu/komplexmicroservice/rendszerservice/api/MenuControllerApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package hu.komplexmicroservice.rendszerservice.api;

import hu.komplexmicroservice.rendszerservice.api.model.TMenu2Dto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-26T18:20:09.809915324+02:00[Europe/Budapest]")
@Validated
@Tag(name = "MenuController", description = "the MenuController API")
public interface MenuControllerApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /DMMenuapi
     *
     * @param userid  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getMenuByUserId",
        tags = { "MenuController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/DMMenuapi/{userid}",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    default ResponseEntity<List<TMenu2Dto>> getMenuByUserId(
        @Parameter(name = "userid", description = "") @Valid @RequestParam(value = "userid", required = false) Long userid
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"rdbcon_postgresuser\" : \"\", \"rdbcon_user\" : \"\", \"rdbcon_password\" : \"\", \"rdbcon_host\" : \"\", \"rdbcon_postgrespassword\" : \"\", \"rdbcon_port\" : \"\", \"rdbcon_id\" : \"\", \"rdbcon_letre_felh_nev\" : \"\", \"rdbcon_title\" : \"\", \"rdbcon_databasename\" : \"\", \"rdbcon_letre_dat\" : \"\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}