# src/main/openapi/rendszer-service-openapi.yml
openapi: 3.1.0
info:
  title: Rendszer-service
  description: Rendszer-service
  version: 1.0.0
servers:
  - url: 'https://localhost:8187'
paths:
   '/DMMenuapi/{userid}':
     get:
      tags:
        - MenuController
      operationId: getMenuByUserId
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TMenu2Dto'

   '/DMMenuparameterek/{menu_id}':
     get:
       tags:
         - MenuParameterekController
       operationId: getParameterekByUserId
       parameters:
         - name: menu_id
           in: path
           required: true
           schema:
             type: Long
       responses:
         '200':
           description: OK
           content:
             application/json:
               schema:
                 type: array
                 items:
                   $ref: '#/components/schemas/TMenuparameterekDto'


components:
  schemas:
    TMenu2Dto:
      type: object
      properties:
        menu_id:
          type: integer
          format: int64
          nullable: false
        menu_gyoker:
          type: integer
          format: int64
          nullable: false
        menu_gyokersorrend:
          type: integer
          format: int64
          nullable: false
        menu_title:
          type: string
          nullable: false
        menu_link:
          type: string
          nullable: false
        menu_parentid:
          type: integer
          format: int64
          nullable: false
        menu_letre_felh_nev:
          type: string
          nullable: false
        menu_letre_dat:
          type: string
          nullable: false
        menu_action:
          type: string
          nullable: false
        menu_sorrend:
          type: integer
          format: int64
          nullable: false
        menu_image_index:
          type: integer
          format: int64
          nullable: false
        menu_shortcut:
          type: string
          nullable: false
        menu_hlevel:
          type: integer
          format: int64
          nullable: false
        szrpfnk_uj_eng:
          type: boolean
          nullable: false
        szrpfnk_modositas_eng:
          type: boolean
          nullable: false
        szrpfnk_visible_engs:
          type: boolean
          nullable: false
    TSearchUserDto:
      type: object
      properties:
        userid:
          type: string
          nullable: false
    TSearchMenuparameterekDto:
      type: object
      properties:
        menu_id:
          type: integer
          format: int64
          nullable: false
    TMenuparameterekDto:
      type: object
      properties:
        menuprm_id:
          type: integer
          format: int64
          nullable: false
        menuprm_menu_id:
          type: integer
          format: int64
          nullable: false
        menuprm_name:
          type: string
          nullable: false
        menuprm_value:
          type: string
          nullable: false
        menuprm_letre_felh_nev:
          type: string
          nullable: false
        menuprm_letre_dat:
          type: string
          nullable: false